requirements.txt
tensorflow>=2.0.0
numpy>=1.19.2
matplotlib>=3.3.2
Pillow>=8.0.0
scikit-learn>=0.24.0